Add navbar and test the navtabs directive
Adding the navbar will show the usage of the $routeProvider in a clearer way.
It also allows us to possibly cover the testing of an angular.js directive.

Because we require some jQuery functions in the test case for the navtabs
directive, we track the jQuery library starting from this patch.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>